+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added CONTRIBUTORS.md
 * Fixes bugs with urls that have dashes in them
 * Fixes bugs with urls containing encoded data (usually images)
+* Deprecates CSSCSS_DEBUG in favor of --show-parser-errors
 
 ## 1.0.0 - 4/7/2013 ##